// .tmp/zip_snapshot/tote-bag-practicantes/apps/api/src/generated/client/enums.ts
export const ShipmentStatus = {
  PENDING: 'PENDING',
  READY_TO_SHIP: 'READY_TO_SHIP',
  SHIPPED: 'SHIPPED',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
  RETURNED: 'RETURNED',
  CANCELLED: 'CANCELLED'
} as const

export type ShipmentStatus = (typeof ShipmentStatus)[keyof typeof ShipmentStatus]
// ...
export const OrderStatus = {
  PENDIENTE_PAGO: 'PENDIENTE_PAGO',
  PAGADA: 'PAGADA',
  EN_PRODUCCION: 'EN_PRODUCCION',
  ENVIADA: 'ENVIADA',
  ENTREGADA: 'ENTREGADA',
  RETURNED_TO_STOCK: 'RETURNED_TO_STOCK',
  CANCELADA: 'CANCELADA'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]